Add files via upload
Heart disease is the leading cause of the death in the United States as like said before. It is also a major cause of disability. There are many things that can raise your risk for heart disease. They are called risk factors. Some of them you cannot control, but there are many that you can control. Learning about them can lower your risk of heart disease.
Our study analysis 13 factors with 200 records of different patients which caused the “Heart-Disease”