Add files via upload
The above analysis is the observation of 157 countries hapiness scores and rank. Study which will help to understand the what are all the factors that increases the happiness score. what is the contribution of GDP per capita and trust on the goverment that helps to improve the Quality of living.